modify logic of updating the max transaction number in a block dynamically
fix param checking related to ledger

fix initialization of m_maxBlockCanSeal && fix code according to codeFactor

+++ libledger/LedgerParam.h
@@ -84,9 +84,9 @@ struct StorageParam
     // for amop storage
     std::string topic;
     size_t timeout;
-    size_t maxRetry;
-    size_t maxStoreKey;
-    size_t maxForwardBlock;
+    int maxRetry;
+    int maxStoreKey;
+    int maxForwardBlock;
 };
 struct StateParam
 {
